all tests for PCA passing
